исправлен тип хранения количества паролей int-> double

+++ GenPassword.h
@@ -21,9 +21,9 @@ public:
   void Init(int len, string& range);
   bool Next();
   virtual void CreatePassword();
-  long CountPasswords();
+  double CountPasswords();
   string  msgErr;
-  long numbPassword;
+  double numbPassword;
 protected:
   char minValue;
   char maxValue;
